Created starting code
Main.py is where the main UI is. Then in pages folder, is where you create your new frames for new pages. In classes folder is where we have our classes that represent objects in the database. In assets folder we have our database file, but also our python file that created the database.