Some fixes and additions
Fixed the header, previously known as nav, so now the buttons and text are now centered. Application now automatically starts off on login screen due running openPage in App's constructor, this was just for development purposes. Created the mark up for the user login page and the registration page, but logic for logging in and registrating users has not been added yet. Next I will do the user account page, and then after I'll get started on the logic for logging in and registering.