Added user account page and database
Finished the mark up for the user account page. Then we finished creating a 'users' table and a database to hold that table. It uses sqlalchemy and orm, which basically means we should be able to manipulate the database with python classes rather than using sql statements. Next up, I'm probably going to work on the mark up for creating other pages, editAccountPage, deleteAccountPage, and a changePasswordPage.