Updates with current pages and additions
All pages with forms now have an option to clear the form. All form fields relating to passwords now have the option to toggle whether the password is hidden or not. Created page for editing accounts, deleting accounts, and changing passwords. Added utilites, which is a file that contains good utility functions for doing aggregate operations on entry widgets. Next is the editAvatarPage.