Basic user registration completed
Successfully implemented orm for not only Users, but also the stories and the messages contained in each story. Successfully added input validation for stuff such as emails, passwords, and usernames. Also added a folder called 'images' in assets. This would be for profile pictures or avatars. Storing the images locally rather than doing http requests for them over the web, which is just a preference on my part. Next functionality should be logging in