Login and logout done
Basic login functionality is done. Logout functionality is done as well, and now we're successfully able to load in the avatar of the user from our images folder. As well as this when you register successfully, it redirects you to the login page, and when you logout, it redirects you to the login page. In both cases, the state of the logged in user changes, and the nav will change accordingly