Fixes and added another page
Did a very basic draft up on the story library page. Just displays stories and we're able to delete the user's stories from the database. Fixed image size in userAccountPage. Then moved to only have one 'session' object throughout the application, which is in 'master' or 'App' class. NOTE: Try to only use the session in master since it's best to only have one session at a time that's interacting with the database