AI settings page and some fixes/changes
Created the basic mark up and gui for the ai settings page. Values of the sliders on that page may change later. The function for logging out the user is now in the 'App' class so that multiple pages can use the logic. This also fixed the bug when user's changed their password, where  it didn't actually log them out  and so the nav bar buttons were still accessible to them.