Finished editing account functionality
Finished functionality for the edit account page. Now the user can edit their accounts, but as well as this, the form fields will be filled with their current account information. This makes it so the user does not have to fill out all of their info when changing their account information, as they can leave some information unchanged.