Added more pages and changes
Fixed clearing entry widgets. Added homePage which allows a user to continue the previous story or do a new story. Added remixStoryPage, which takes an existing story and the user's input to remix a story.Finally we added saveStoryPage which lets a user save the story from their most recent session. Just added the GUI, not much back-end application stuff yet due to how we're still waiting on ai portion and chat page portion. However, we added an array in master that keeps track of the messages in the user's current session. Should help allow the user to continue their previous story with an ai, without having to commit to saving their story.