AI chat page and successful message history tracker
Created a very basic ai chat page that has a gui, but no ai has been integrated; will be modified when the real one is implemented. Created logic for tracking the messages for the current story that the user is writing. Home page now is able to link to ai chat page successfully, however it's still unfinished as there needs to be conditions for being able to continue a previous story. Also fixed the storyLibraryPage for rendering story cards, parts of the cards weren't on screen, but now it's fixed. Both remixStoryPage and saveStoryPage are still in infancy and logic hasn't been tested yet, so that's going to be the goal for the next commit.