Saving and rendering updated
Library page successfully redirects user to the ai chat page when they want to continue the story, and correctly renders the previous messages in their saved story. Save story page is able to dynamically change layout and form depending on whether the user is saving changes to an existing story or saving a completely new story. The saving functionality seems to work well as well. homePage is able to start a fresh new chat if the user wants to start a new story, and it's able to allow the user to continue a story where they left off as it successfully renders the messages. This works in both cases of the user continuing a saved story, or if the user is creating a new story but hasn't saved it yet. This is possible because we are storing the unsaved messages that the user hasn't saved yet, so regardless if the user is currently continuing a saved story or not, the unsaved messages that the user put in their recent session will still render