Finished basic storyRemixPage and its logic
Finished the logic and functionality for how we remix stories and how to persist and handle those related story messages. The code file is becoming unruly, so in next commit I'm going to separate the pages into files, and then try to find a better way to implement how we can know when we're continuing a saved story, remixing a story, etc. Because right now, it feels somewhat messy to handle two different variables for holding story objects.