Updated story system and modularized code
Updated system on how we handle story objects, and how we handle when the user is continuing a saved story or remixing an existing one. Also modularized the code so that the different pages are in their own file. The 'Main.py' file was getting too big, and this is much more readable and maintainable.