History Manager Changes, Response Overhaul
Started on some history manager changes, added the `clear` and `populate` methods to the list. The major change is the way that `complete` is used. It is now an embedded call, only being called in the class rather than by the developer. This means developers using this program only have to make one call to get a response.