Set up AI and GUI
Set up AI class instance in App class, so now storyGPT should be available across the application. AISettingsPage is done as you should be able to change the settings of storyGPT through it. For ai.py, storyGPT now has instance attributes for the response length and writing style, which makes it a lot easier for the settings page. As well as this, set up the code to read an openai api key from a file 'api_key.txt' in the 'assets' folder. Finally, added some utility functions that will be useful in further integrating the AI to our gui application.