Linked AI to more application pages
Added appropriate AI logic to the homePage, storyLibraryPage, remixStoryPage, and AIChatPage. For the AIChatPage, in order to prevent potential issue of the user leaving the page as the AI generated content I implemented a solution where the user was couldn't change pages until the AI content was finished generating. Also created logic so that AI message chunks would be rendered in real-time, but I haven't and I currently am not able to test this out with openai yet.