Changes with organizing and AISettingsPage
Changed it so that the generator returned by the AI would be stored in one variable called storyGenObj, which is located in 'App' class. Changed the AISettingsPage so that stream is no longer an option. For other files, did some minor spacing changes or type ignore, but other than that no functionality changes.