Dark mode finished
Theme toggling is finished, however everytime themes are switched, the state of the page is reset. Most of this has been dealt with, but the big issue is how form data is cleared when toggling themes due to how we re-create/re-render the tkinter-frame. If possible, next commit is going to work on a way so that allows us to save form data on input change and use it if needed.