Majorly improved story exportation
Users are now able to do 'save as' for their pdfs and for the bulk zip files. This not only makes saving more user friendly, but also it makes it so the application will let the user know when a file that's been saved already exists in a directory, when before if a file with the same name existed the application would do nothing.