Fixed some errors in the AI, added user token and changed app logo
Fixed some errors in the AI that would happen if a user typed whitespaces before the message triggering an Open AI exception causing the app to halt. Additionally, I added logic for a login user token that remembers the previously logged in user until a user logs out. Next, I changed our app icon from the worm to the book as our team agreed to.