Refactored the ModelBase Class and Other small changes
Added a ChatHistoryManager class and moved all my chat history management methods from ModelBase to ChatHistoryManager. This not only makes my code cleaner, but it adheres to the Single Responsibility Principle.

I moved some code from the init function in AIChatPage to a setup method to make the code more readable. I also slightly changed the way that the program renders the moving of the text bubbles to make it more smooth.